Revert "feat: add post index create destroy etc..　#3"
