+++ .gitignore
@@ -23,9 +23,7 @@
 !/storage/.keep
 
 /public/assets
-public/uploads/
 .byebug_history
-
 
 # Ignore master key for decrypting credentials and more.
 /config/master.key
