+++ app/controllers/application_controller.rb
@@ -1,13 +1,3 @@
 class ApplicationController < ActionController::Base
   include SessionsHelper
-
-  private
-
-  def logged_in_user
-    unless logged_in?
-      store_location
-      flash[:danger] = "ログインしてください"
-      redirect_to login_url
-    end
-  end
 end
