+++ app/controllers/post_pages_controller.rb
@@ -1,0 +1,4 @@
+class PostPagesController < ApplicationController
+  def home
+  end
+end
