+++ app/controllers/users_controller.rb
@@ -1,16 +1,14 @@
 class UsersController < ApplicationController
   before_action :logged_in_user, only: [:index, :edit, :update, :destroy]
   before_action :correct_user,   only: [:edit, :update]
-  before_action :admin_user,     only: [:destroy]
+  before_action :admin_user,     only: :destroy
 
   def index
     @users = User.all
-    @users = User.page(params[:page]).per(20)
   end
 
   def show
     @user = User.find(params[:id])
-    @posts = @user.posts.page(params[:page])
   end
 
   def new
@@ -48,6 +46,17 @@ class UsersController < ApplicationController
     redirect_to users_url
   end
 
+  # def guest
+  #   if logged_in?
+  #     flash[:notice] = 'すでにログインしています'
+  #   else
+  #     guest_user = User.find_by(email: "guest@guest.com")
+  #     log_in(guest_user)
+  #     flash[:notice] = 'ゲストユーザーとしてログインしました'
+  #   end
+  #     redirect_to root_path
+  # end
+
   def guest
     user = User.find_or_create_by!(email: 'guest@example.com') do |user|
       user.name = "guestuser"
@@ -61,6 +70,14 @@ class UsersController < ApplicationController
 
     def user_params
       params.require(:user).permit(:name, :email, :password, :password_confirmation)
+    end
+
+    def logged_in_user
+      unless logged_in?
+        store_location
+        flash[:danger] = "ログインしてください"
+        redirect_to login_url
+      end
     end
 
     def correct_user
