+++ app/helpers/post_pages_helper.rb
@@ -1,0 +1,2 @@
+module PostPagesHelper
+end
