--- app/helpers/posts_helper.rb
@@ -1,2 +1,0 @@
-module PostsHelper
-end
