+++ app/javascript/packs/application.js
@@ -2,7 +2,8 @@
 // present in this directory. You're encouraged to place your actual application logic in
 // a relevant structure within app/javascript and only use these pack files to reference
 // that code so it'll be compiled.
-//= require rails-ujs
+//= jquery
+//= jquery_ujs
 
 import Rails from "@rails/ujs"
 import Turbolinks from "turbolinks"
@@ -13,4 +14,5 @@ Rails.start()
 Turbolinks.start()
 ActiveStorage.start()
 
+import "stylesheets/application"
 import '../css/tailwind.css';
