--- app/uploaders/image_uploader.rb
@@ -1,59 +1,0 @@
-class ImageUploader < CarrierWave::Uploader::Base
-  # Include RMagick or MiniMagick support:
-  # include CarrierWave::RMagick
-  include CarrierWave::MiniMagick
-
-  # Choose what kind of storage to use for this uploader:
-  storage :file
-  # storage :fog
-
-  # Override the directory where uploaded files will be stored.
-  # This is a sensible default for uploaders that are meant to be mounted:
-  def store_dir
-    "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
-  end
-
-  # Provide a default URL as a default if there hasn't been a file uploaded:
-  # def default_url(*args)
-  #   # For Rails 3.1+ asset pipeline compatibility:
-  #   # ActionController::Base.helpers.asset_path("fallback/" + [version_name, "default.png"].compact.join('_'))
-  #
-  #   "/images/fallback/" + [version_name, "default.png"].compact.join('_')
-  # end
-
-  # Process files as they are uploaded:
-  # process scale: [200, 300]
-  # process :convert => 'jpg'
-  #
-  # def scale(width, height)
-  #   # do something
-  # end
-  # process :resize_to_limit => [700, 700]
-
-  # Create different versions of your uploaded files:
-
-  version :small do
-    process resize_to_fill: [300, 300]
-  end
-
-  # Add a white list of extensions which are allowed to be uploaded.
-  # For images you might use something like this:
-  def extension_whitelist
-    %w(jpg jpeg gif png)
-  end
-
-  # Override the filename of the uploaded files:
-  # Avoid using model.id or version_name here, see uploader/store.rb for details.
-  # def filename
-  #   super.chomp(File.extname(super)) + '.jpg'
-  # end
-
-  # def filename
-  #   if original_filename.present?
-  #     time = Time.now
-  #     name = time.strftime('%Y%m%d%H%M%S') + '.jpg'
-  #     name.downcase
-  #   end
-  # end
-
-end
