--- app/views/posts/_post.html.erb
@@ -1,11 +1,0 @@
-<li post-<%= post.id %>>
-  <%= link_to post.user.name, post.user %>
-  <%= post.content %>
-  <%= image_tag post.image.url(:small) %>
-
-<p>投稿記事 <%= time_ago_in_words(post.created_at) %> 前 </p>
-  <% if current_user?(post.user) %>
-    <%= link_to   "編集", edit_post_path(post) %>
-    <%= button_to "削除", post_path(post), method: :delete %>
-  <% end %>
-</li>
