--- app/views/posts/edit.html.erb
@@ -1,19 +1,0 @@
-<h1>Posts#edit</h1>
-  <p>現在ログイン中のユーザー：<%= current_user.name %></p>
-
-<%= form_with(model: @post, local: true) do |f| %>
-
-<%= render 'shared/error_messages', object: f.object %>
-
-  <div>
-    <%= f.label :中身 %><br>
-    <%= f.text_field :content, autofocus: true, :placeholder =>"#{@post.content}" %>
-  </div>
-
-  <div>
-    <%= f.label :画像 %>
-    <%= f.file_field :image, :placeholder =>"#{@post.image}" %>
-  </div>
-
-  <div><%= f.submit "更新する" %></div>
-<% end %>
