--- app/views/posts/show.html.erb
@@ -1,8 +1,0 @@
-<div >
-  <p><%= @post.content %></p>
-  <%= image_tag @post.image.url(:small) %>
-</div>
-
-<div>
-  <p><%= @post.user.name %></p>
-</div>
