+++ app/views/sessions/new.html.erb
@@ -1,8 +1,9 @@
 <h1>ログイン</h1>
 
-<div>
-  <div>
+<div class="row">
+  <div class="col-md-6 col-md-offset-3">
     <%= form_with(url: login_path, scope: :session, local: true) do |f| %>
+
       <%= f.label :メールアドレス %>
       <%= f.email_field :email, class: 'form-control' %>
 
@@ -11,7 +12,7 @@
 
       <%= f.label :remember_me, class: "checkbox inline" do %>
         <%= f.check_box :remember_me %>
-        <p>次回から自動でログイン</p>
+        <span>次回から自動でログイン</span>
       <% end %>
 
       <%= f.submit "ログイン", class: "btn btn-primary" %>
