+++ app/views/shared/_error_messages.html.erb
@@ -1,10 +1,10 @@
-<% if object.errors.any? %>
-  <div>
-    <div>
-      The form contains<%= pluralize(object.errors.count, "error") %>.
+<% if @user.errors.any? %>
+  <div class="alert alert-warning">
+    <div class="alert alert-danger">
+      The form contains <%= pluralize(@user.errors.count, "error") %>.
     </div>
     <ul>
-      <% object.errors.full_messages.each do |message| %>
+      <% @user.errors.full_messages.each do |message| %>
         <li><%= message %></li>
       <% end %>
     </ul>
