+++ app/views/shared/_header.html.erb
@@ -1,20 +1,29 @@
 
 <header class="navbar navbar-fixed-top navbar-inverse">
   <div class="container">
-    <%= link_to "Remoshare", root_url, id: "logo" %>
+    <%= link_to "Remoshare", root_path, id: "logo" %>
     <nav>
-      <ul>
-          <li><%= link_to "投稿一覧", root_path %></li>
-          <li><%= link_to "チャットで相談", '#' %></li>
+      <ul class="nav navbar-nav navbar-right">
+        <li><%= link_to "投稿一覧", root_path %></li>
+        <li><%= link_to "チャットで相談", '#' %></li>
 
         <% if logged_in? %>
-          <li><%= link_to "投稿", new_post_path %></li>
           <li><%= link_to "ユーザー一覧", users_path %></li>
           <li><%= link_to "通知", '#' %></li>
-          <li><%= link_to "プロフィール", current_user %></li>
-          <li><%= link_to "設定", edit_user_path(current_user) %></li>
-          <li><%= button_to "ログアウト", logout_path, method: :delete %></li>
-        <% else %>
+          <li class="dropdown">
+            <a href="#" class="dropdown-toggle" data-toggle="dropdown">
+              アカウント <b class="caret"></b>
+            </a>
+            <ul class="dropdown-menu">
+              <li><%= link_to "プロフィール", current_user %></li>
+              <li><%= link_to "設定", edit_user_path(current_user) %></li>
+              <li class="divider"></li>
+              <li>
+                <%= button_to "ログアウト", logout_path, method: :delete %>
+              </li>
+            </ul>
+          </li>
+          <% else %>
           <li><%= link_to "ログイン", login_path %></li>
         <% end %>
       </ul>
