+++ app/views/users/index.html.erb
@@ -1,7 +1,13 @@
-<h1>ユーザー一覧</h1>
+<h1>All users</h1>
 
 <ul class="users">
-  <%= render @users %>
+  <% @users.each do |user| %>
+    <li>
+      <%= link_to user.name, user %>
+      <% if current_user.admin? && !current_user?(user) %>
+      | <%= button_to "delete", user, method: :delete,
+                                  data: { confirm: "本当に削除しますか?" } %>
+  <% end %>
+    </li>
+  <% end %>
 </ul>
-
-<%= paginate @users%>
