--- config/initializers/time_formats.rb
@@ -1,1 +1,0 @@
-Time::DATE_FORMATS[:datetime_base] = '%Y/%m/%d %H:%M'
