+++ config/webpack/environment.js
@@ -1,11 +1,11 @@
 const { environment } = require('@rails/webpacker')
 
-// const webpack = require('webpack')
-// environment.plugins.prepend('Provide',
-//   new webpack.ProvidePlugin({
-//     $: 'jquery/src/jquery',
-//     jQuery: 'jquery/src/jquery'
-//   })
-// )
+const webpack = require('webpack')
+environment.plugins.prepend('Provide',
+  new webpack.ProvidePlugin({
+    $: 'jquery/src/jquery',
+    jQuery: 'jquery/src/jquery'
+  })
+)
 
 module.exports = environment
