--- db/migrate/20210213090555_create_posts.rb
@@ -1,12 +1,0 @@
-class CreatePosts < ActiveRecord::Migration[6.1]
-  def change
-    create_table :posts do |t|
-      t.text :content, null: false
-      t.string :image, null: false
-      t.references :user, foreign_key: true
-
-      t.timestamps
-    end
-    add_index :posts, [:user_id, :created_at]
-  end
-end
