+++ db/seeds.rb
@@ -1,16 +1,7 @@
-# # ユーザー作成
-# 1.upto(39) do |i|
-#   name  = Faker::Name.name
-#   email = "sample-#{i}@example.com"
-#   password = 'password'
-#   User.create!(name: name,
-#                 email: email,
-#                 password: password,
-#                 password_confirmation: password)
-# end
-
-# users = User.order(:created_at).take(6)
-# 30.times do
-#   content = Faker::Lorem.sentence(word_count: 5)
-#   users.each { |user| user.posts.create!(content: content) }
-# end
+# This file should contain all the record creation needed to seed the database with its default values.
+# The data can then be loaded with the bin/rails db:seed command (or created alongside the database with db:setup).
+#
+# Examples:
+#
+#   movies = Movie.create([{ name: 'Star Wars' }, { name: 'Lord of the Rings' }])
+#   Character.create(name: 'Luke', movie: movies.first)
