--- test/fixtures/posts.yml
@@ -1,11 +1,0 @@
-# Read about fixtures at https://api.rubyonrails.org/classes/ActiveRecord/FixtureSet.html
-
-one:
-  content: MyText
-  image: MyString
-  user: one
-
-two:
-  content: MyText
-  image: MyString
-  user: two
